pyproject.toml
[project]
name = "wechat-robot-mcp-server"
version = "1.0.0"
description = "WeChat Robot MCP Server - Python Implementation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
    "aiohttp>=3.9.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
]

[project.scripts]
wechat-robot-mcp-server = "src.main:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
